+++ onlyofficeeditorapi.php
@@ -30,6 +30,20 @@ $apitype = required_param('apiType', PARAM_TEXT);
 
 $context = CONTEXT_MODULE::instance($cmid);
 switch ($apitype) {
+    case 'mention':
+        require_capability('mod/onlyofficeeditor:editdocument', $context);
+        try {
+            $courseid = $_GET['courseid'];
+            require_login($courseid);
+            $actionlink = $_POST['link'];
+            $emails = $_POST['emails'];
+            $comment = $_POST['comment'];
+            \mod_onlyofficeeditor\util::mention_user_in_comment($actionlink, $comment, $emails, $context);
+            echo json_encode($comment);
+        } catch (moodle_exception $e) {
+            throw new \Exception();
+        }
+        break;
     case 'sections':
         try {
             $courseid = $_GET['courseid'];